Move CSV codec into a separate module
